+++ crates/brynja-crypto-cpu-std/src/protected_memory/unsupported.rs
@@ -4,6 +4,12 @@ use super::Error;
 pub(super) enum Mapping {}
 
 impl Mapping {
+    pub(super) fn stack(_: usize, _: usize) -> Result<Self, Error> {
+        Err(Error::Unsupported)
+    }
+    pub(super) fn run<F: FnOnce() + Send>(&mut self, _: F) -> Result<(), Error> {
+        match *self {}
+    }
     pub(super) fn new(_: usize, _: usize) -> Result<Self, Error> {
         Err(Error::Unsupported)
     }